+++ packages/go/package.json
@@ -12,7 +12,7 @@
   "scripts": {
     "build": "node ../../utils/build-builder.mjs",
     "test": "jest --reporters=default --reporters=jest-junit --env node --verbose --runInBand --bail",
-    "test-e2e": "pnpm test",
+    "test-e2e": "pnpm test test/integration-*",
     "type-check": "tsc --noEmit"
   },
   "files": [
