--- packages/go/test/fixtures.test.js
@@ -1,18 +1,0 @@
-const fs = require('fs');
-const path = require('path');
-
-const {
-  testDeployment,
-} = require('../../../test/lib/deployment/test-deployment.js');
-
-jest.setTimeout(4 * 60 * 1000);
-
-const fixturesPath = path.resolve(__dirname, 'fixtures');
-
-for (const fixture of fs.readdirSync(fixturesPath)) {
-  it(`should build ${fixture}`, async () => {
-    await expect(
-      testDeployment(path.join(fixturesPath, fixture))
-    ).resolves.toBeDefined();
-  });
-}
