+++ packages/go/test/fixtures/29-gin-server-go1.26/.gitignore
@@ -1,0 +1,3 @@
+.gocache/
+.gomodcache/
+.gopath/
