+++ packages/go/test/fixtures/30-net-http-server-go1.26/.gitignore
@@ -1,0 +1,3 @@
+.gocache/
+.gomodcache/
+.gopath/
