+++ packages/go/test/fixtures/30-net-http-server-go1.26/go.mod
@@ -1,3 +1,3 @@
-module serverful-net-http-go126
+module net-http-server-go126
 
 go 1.26
