+++ packages/go/test/fixtures/31-chi-server-go1.26/.gitignore
@@ -1,0 +1,3 @@
+.gocache/
+.gomodcache/
+.gopath/
