+++ packages/go/test/fixtures/31-chi-server-go1.26/go.mod
@@ -1,4 +1,4 @@
-module serverful-chi-go126
+module chi-server-go126
 
 go 1.26
