+++ packages/go/test/fixtures/32-net-http-server-go1.25/.gitignore
@@ -1,0 +1,3 @@
+.gocache/
+.gomodcache/
+.gopath/
