+++ packages/go/test/fixtures/32-net-http-server-go1.25/go.mod
@@ -1,3 +1,3 @@
-module serverful-net-http-go125
+module net-http-server-go125
 
 go 1.25
