+++ packages/go/test/integration-1.test.js
@@ -1,0 +1,1 @@
+require('./integration-setup')(1);
